+++ public/scripts/src/store/plugins.js
@@ -1,5 +1,11 @@
-import wsplugin from '../wsplugin';
+import login from '../wsplugins/login';
+import ms from '../wsplugins/ms';
+import game from '../wsplugins/game';
+
+const socket = io(window.location.protocol + '//' + window.location.host);
 
 export default [
-	wsplugin(io(window.location.protocol + '//' + window.location.host))
+	login(socket),
+	ms(socket),
+	game(socket)
 ];
